celox-common: Enable MTP and ADB by default
On a cleanly flashed system, neither MTP or UMS are selected and ADB is enabled. In this configuration:

persist.sys.usb.config=adb

ADB will not work and computers are unable to mount the devices storage. This is confusing to users who are usually unaware of the options in the storage section which will enable either MTP or UMS. Once MTP/UMS is enabled:

persist.sys.usb.config=mtp,adb

then ADB starts working and computers will recognize the devices storage.

adding this to the build.prop stops confusion for new users as the device should now be recognized out of the box and this also allows logcat to be pull while the device is starting up for debugging purposes. Both are assumed to be enabled by default by end users anyways, so they won't even notice this ;-)